# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "API de Configuração",
        "description": "Essa API tem como objectivo configurar a documentação com o Swagger",
        "contact": {
            "email": "gerson.01@outlook.pt"
        },
        "version": "1.0.0"
    },
    "tags": [
        {
            "name": "provincias"
        },
        {
            "name": "escola"
        }
    ],
    "paths": {
        "/provincias": {
            "get": {
                "description": "Retorna as Provincias",
                "tags": ["provincias"],
                "parameters": [],
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref":"#/definitions/Provincias"
                        }
                    },
                    "400": {
                        "description": "Nenhuma província Listada. Carrega o ficheiro JSON das províncias usando o POST"
                    },
                    "500": {
                        "description": "Erro do Servidor"
                    }
                }
            },
            "post": {
                "description": "Carrega o JSON das Provincias",
                "tags": ["provincias"],
                "consumes": ["multipart/form-data"],
                "parameters":[
                    {
                        "in": "formData",
                        "name": "file",
                        "type": "file",
                        "description": "JSON das provincias"
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "$ref":"#/definitions/Provincias"
                        }
                    },
                    "400": {
                        "description": "Nenhuma província Listada. Carrega o ficheiro JSON das províncias usando o POST"
                    },
                    "500": {
                        "description": "Erro do Servidor"
                    }
                }
            }
        },
        "/escola": {
            "get": {
                "description": "Retorna todas Escolas",
                "tags": ["escola"],
                "parameters": [],
                "produces": ["application/json"],
                "responses": {
                    "400": {
                        "description": "Nenhuma escola Listada"
                    },
                    "200": {
                        "schema": {
                            "$ref":"#/definitions/Escola"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria a escola através de parâmetros",
                "tags": ["escola"],
                "consumes": ["application/json"],
                "produces":["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Parametros da Escola",
                    "required": "true",
                    "schema": {
                        "$ref": "#/definitions/Escola"
                    }
                }],
                "responses": {
                    "400": {
                        "description": "Por favor digite a Província da forma certa" 
                    },
                    "200": {
                        "schema": {
                            "$ref":"#/definitions/Escola"
                        }
                    }
                }
            }
        },
        "/escola/{id}": {
            "get":{
                "description":"Buscar escola por ID",
                "tags":["escola"],
                "produces":["application/json"],
                "parameters":[{
                    "name":"id",
                    "in":"path",
                    "description":"Retorna escola por ID",
                    "required":"true",
                    "type":"string"
                }],
                "responses":{
                    "400":{
                        "description":"Nenhuma escola Listada."
                    },
                    "200":{
                        "schema":{
                            "$ref":"#/definitions/Escola"
                        }
                    },
                    "500":{
                        "description":"Erro do Servidor"
                    }
                }
            },
            "patch": {
                "description":"Actualização da Escola",
                "tags":["escola"],
                "parameters":
                [
                    {
                        "name":"id",
                        "in":"path",
                        "description":"Retorna escola por ID",
                        "required":"true",
                        "type":"string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Parametros da Escola",
                        "required": "true",
                        "schema": {
                        "$ref": "#/definitions/Escola"
                        }
                    }
                ],
                "responses":{
                    "400":{
                        "description":"Preencha os campos nome, email, número de salas e província da escola"
                    },
                    "401":{
                        "description":"Por favor digite a Província da forma certa"
                    },
                    "200":{
                        "schema":{
                            "$ref":"#/definitions/Escola"
                        }
                    }
                }
            },
            "delete":{
                "description":"Deletar escola por ID",
                "tags":["escola"],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"Retorna escola por ID",
                        "required":"true",
                        "type":"string"
                    }
                ],
                "responses":{
                    "404":{
                        "description":"Escola não encontrada"
                    },
                    "200":{
                        "description":"Escola deletada"
                    }
                }
            }
        },
        "/escola/excel":{
            "post": {
                "description": "Carrega o EXCEL com as Escolas",
                "tags": ["escola"],
                "consumes": ["multipart/form-data"],
                "parameters":[
                    {
                        "in": "formData",
                        "name": "file",
                        "type": "file",
                        "description": "O EXCEL deve ter as colunas na ordem Nome, Email, Número de Salas, Província"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Nenhum EXCEL carregado"
                    },
                    "401": {
                        "description": "Erro de Escrita da Província no EXCEL. Esse erro poe a App em cashed."
                    },
                    "200": {
                        "schema":{
                            "$ref":"#/definitions/Escola"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Provincias":{
            "type":"object",
            "properties":{
                "id":{
                    "type":"string"
                },
                "nome":{
                    "type":"string"
                }
            }
        },
        "Escola": {
            "type": "object",
            "properties": {
                "nome": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "numSalas": {
                    "type": "number"
                },
                "provincia": {
                    "type": "string"
                }
            },
            "required": ["nome", "email", "numSalas", "provincia"]
        }
    }
}